fixed validation in actions and tests

+++ app/Actions/User/UserDeleteByIdAction.php
@@ -10,10 +10,14 @@ class UserDeleteByIdAction {
     public function handle(int $id) {
 
         $user = User::find($id);
-        if ($user->avatar) {
-            Storage::delete($user->avatar);
-        }
-        $user->delete();
-        return $user;
+
+        if ($user != null) {
+            if ($user->avatar) {
+                Storage::delete($user->avatar);
+            }
+            $user->delete();
+            return $user;
+        } else
+        return response('Пользователь не найден', 204);
     }
 }
